Move LLMOutputType to cogt
